# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BotStatus = typeof BotStatus[keyof typeof BotStatus];


export const BotStatus = {
  stopped: 'stopped',
  launching: 'launching',
  running: 'running',
  awaiting_auth: 'awaiting_auth',
  error: 'error',
} as const;

export interface Bot {
  id: number;
  name: string;
  accountIndex: number;
  gfnEmail: string | null;
  gfnPassword: string | null;
  serverName: string | null;
  /** Path to the saved browser user-data directory for this bot's session (e.g. C:\EdgeProfiles\Bot01) */
  browserProfilePath: string | null;
  status: BotStatus;
  /** One-time auth code entered by the operator — read by the automation script to complete login */
  authCode: string | null;
  /** When the automation last set the bot to awaiting_auth */
  authCodeRequestedAt: string | null;
  lastActivity: string | null;
  errorMessage: string | null;
}

export interface CreateBotInput {
  name: string;
  accountIndex: number;
  gfnEmail?: string | null;
  gfnPassword?: string | null;
  serverName?: string | null;
  browserProfilePath?: string | null;
}

export interface UpdateBotInput {
  name?: string;
  accountIndex?: number;
  gfnEmail?: string | null;
  gfnPassword?: string | null;
  serverName?: string | null;
  browserProfilePath?: string | null;
  authCode?: string | null;
  status?: BotStatus;
}

export interface BulkActionResult {
  affected: number;
  bots: Bot[];
}