# lib/api-zod/src/generated/types/botStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type BotStatus = typeof BotStatus[keyof typeof BotStatus];


export const BotStatus = {
  stopped: 'stopped',
  launching: 'launching',
  running: 'running',
  awaiting_auth: 'awaiting_auth',
  error: 'error',
} as const;