# lib/api-zod/src/generated/types/bulkActionResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Bot } from './bot';

export interface BulkActionResult {
  affected: number;
  bots: Bot[];
}