# lib/api-zod/src/generated/types/updateBotInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BotStatus } from './botStatus';

export interface UpdateBotInput {
  name?: string;
  accountIndex?: number;
  gfnEmail?: string | null;
  gfnPassword?: string | null;
  serverName?: string | null;
  browserProfilePath?: string | null;
  authCode?: string | null;
  status?: BotStatus;
}